+++ src/paint.js
@@ -1,20 +1,20 @@
 import d3 from './scripts/d3.min';
-import {
-  pageExtensionData
-} from './scripts/irregularUtils';
+import setupPageExtensionData from './pageExtensionData';
+import './extendD3WithLasso';
 
-function setupPaint({ $, lasso, qlik }) {
+function setupPaint({ $, qlik }) {
+  const pageExtensionData = setupPageExtensionData({ $ });
+
   return function ($element, layout) {
     // Call irregularUtils to page the data for > 10000 points
     const maxPages = qlik.navigation.getMode() === "analysis" ? 10 : 1;
-    pageExtensionData(this, $element, layout, heatMap, [qlik, d3, lasso], maxPages);
+    pageExtensionData(this, $element, layout, heatMap, [qlik, d3], maxPages);
 
     function heatMap($element, layout, fullMatrix, _this, ref) {
       $element.empty();
 
       var qlik = ref[0],
-        d3 = ref[1],
-        lasso = [2];
+        d3 = ref[1];
       var app = qlik.currApp();
 
       // get qMatrix data array
